fix(state): extract all types to separate folder
