+++ packages/form-state-manager/src/files/form-manager-context.ts
@@ -1,21 +1,5 @@
-import { createContext, FormEvent } from 'react';
-import AnyObject from '../types/any-object';
-
-export interface Action extends AnyObject {
-  type: string;
-}
-
-export interface ManagerContextValue {
-  values: AnyObject;
-  dispatch: (action: Action) => void;
-  handleSubmit: (event: FormEvent) => void;
-  registerField: (dispatch: (action: Action) => void, fieldState: AnyObject) => void;
-  unRegisterField: (dispatch: (action: Action) => void, fieldState: AnyObject) => void;
-}
-
-export interface ManagerContext {
-  value: ManagerContextValue;
-}
+import { createContext } from 'react';
+import { ManagerContextValue } from '../types/form-manager-context';
 
 const noop = () => {
   /**
