+++ packages/form-state-manager/src/types/form-state-manager.d.ts
@@ -1,0 +1,8 @@
+import AnyObject from "./any-object";
+
+export interface FormStateManagerProps {
+    onSubmit: (values: AnyObject) => void;
+    children: ((props: AnyObject) => React.ReactNode);
+}
+
+export default FormStateManagerProps;
