+++ packages/form-state-manager/src/types/state-manager-reducer.d.ts
@@ -1,0 +1,8 @@
+import AnyObject from "./any-object";
+
+export interface FieldState {
+    name: string;
+    getFieldState: () => AnyObject;
+    value: any;
+    persistOnUnmount: boolean;
+}
