+++ packages/form-state-manager/src/types/use-field.d.ts
@@ -1,0 +1,7 @@
+import AnyObject from './any-object';
+
+export interface UseField extends AnyObject {
+    name: string;
+}
+
+export default UseField;
