+++ packages/form-state-manager/src/types/use-subscription.d.ts
@@ -1,0 +1,11 @@
+import AnyObject from "./any-object";
+
+export type OnChangeEvent = React.ChangeEvent | any;
+
+export interface UseSubscription {
+    name: string;
+    initialValue?: any;
+    subscription?: AnyObject;
+}
+
+export default UseSubscription;
