+++ packages/form-state-manager/src/utils/use-subscription.ts
@@ -1,9 +1,6 @@
 import { useEffect, useState, useContext } from 'react';
-import FormManagerContext, { ManagerContextValue } from '../files/form-manager-context';
-import AnyObject from '../types/any-object';
-
-// tslint:disable-next-line: no-any-union
-type OnChangeEvent = React.ChangeEvent | any;
+import FormManagerContext from '../files/form-manager-context';
+import UseSubscription, { OnChangeEvent } from '../types/use-subscription';
 
 class FieldState {
   constructor(public value: any) {
@@ -32,12 +29,6 @@ const sanitizeValue = (event: OnChangeEvent) => {
 
   return event;
 };
-
-interface UseSubscription {
-  name: string;
-  initialValue?: any;
-  subscription?: AnyObject;
-}
 
 const useSubscription = ({ name, initialValue, subscription = {} }: UseSubscription) => {
   const { registerField, unRegisterField, dispatch } = useContext(FormManagerContext);
